requirements.txt
statsmodels 
yfinance 
apache-airflow-providers-snowflake 
snowflake-connector-python
dbt-snowflake